LogoFeaturePkg,BoardModulePkg: Set existing logo as BGRT
In MinPlatform, FSP-S displays the logo to the user early. However, this
means that no BGRT is written either. To reduce duplication in the flash
image, a new LogoDxe will set this BMP as the BGRT buffer. The previous
LogoDxe drivers here, along with a BootLogoEnableLogo() patch to BDS
libraries, would begin displaying the Intel logo over the board FDF's
choice - presently, these are the TianoCore logo.

Also begin using BootLogoUpdateProgress().

Tested on KabylakeOpenBoardPkg, the boot logo becomes also a BGRT logo.

A follow-up patch can be considered to set all platforms to use Intel's
logo.

Cc: Sai Chaganty <rangasai.v.chaganty@intel.com>
Cc: Isaac Oram <isaac.w.oram@intel.com>
Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
Cc: Ankit Sinha <ankit.sinha@intel.com>
Cc: Eric Dong <eric.dong@intel.com>
Cc: Dandan Bi <dandan.bi@intel.com>
Cc: Liming Gao <gaoliming@byosoft.com.cn>
Signed-off-by: Benjamin Doron <benjamin.doron00@gmail.com>