MinPlatformPkg,WhiskeylakeOpenBoardPkg/SecFspWrapperPlatformSecLib: First BoardInitLib
SecBoardInitLib is called to enable serial port before
SerialPortInitialize and DEBUG().

This is strongly assumed to be necessary for I2cHdmiDebugSerialPortLib
in SEC phase, which presently initialises this way.

No testing was performed before, it's assumed too risky unless the GPIO
happens to have the desired default. Presently, DEBUG() works in SEC
phase.

Signed-off-by: Benjamin Doron <benjamin.doron00@gmail.com>