[WIP] KabylakeOpenBoardPkg/I2cHdmiDebugSerialPortLib: Commit local
While the key patches here should probably be merged into the initial
patches for this library, it appears it isn't being merged soon.
Therefore, commit the patches that improve the library for S3 use.

Other than edits to INF LibraryClasses and header `#include`s, the
primary patch here assists runtime by creating events to toggle a
boolean such that gBS is not used after end-of-BS. This is necessary for
DebugLibSerialPort, but not RSC, which uninstalls the serial port
handler at end-of-BS.

For S3 resume, a key finding was that this is still insufficient. The
image is copied into the lockbox for its own security at
DxeSmmReadyToLock, so toggling booleans in the data section is
ineffective. Early DXE is fairly single-threaded and testing indicates
that simply consuming ..TplNull.c is workable.

Also, GCC 12 requires a patch to a `switch` block that
improved flow coherency.

Signed-off-by: Benjamin Doron <benjamin.doron00@gmail.com>