EarlySpiFlashRescueFeaturePkg: Initial commit of new feature
Implement a new debug feature to assist in recovering the SPI flash from
broken firmware. This feature is intended to be used with the userspace
tool presently in my repo here:
https://github.com/benjamindoron/early_flash_rescue.

After my planned second project for GSoC 2022 finished early, I chose to
implement this feature too with the remaining time.

This PEIM can load successfully in either CAR or DRAM. Further testing
required to confirm that it works after resolving a bug where `DEBUG()`
seemed to overrun the transport layer's internal buffer, and commands
were lost. The UEFI shell app for testing in my repo works well.

Speed of this feature depends on the SerialPortLib implementation. While
PCDs ensure the PEIM is implementation-agnostic, two `MicroSecondDelay()`
calls are targeted at the Bus Pirate transport for the GMBUS
SerialPortLib.

With that SerialPortLib and transport, time taken ranges from ~5
minutes to flash 70 blocks, to ~30 minutes or more to flash MinPlatform
when code changes modify many blocks. It is estimated that 1h40m is
required to flash all blocks of 6M BIOS region, but this is unlikely:
MinPlatform uses less space. Nonetheless, optimisation is still desired.

Cc: Sai Chaganty <rangasai.v.chaganty@intel.com>
Cc: Isaac Oram <isaac.w.oram@intel.com>
Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
Cc: Ankit Sinha <ankit.sinha@intel.com>
Cc: Liming Gao <gaoliming@byosoft.com.cn>
Cc: Eric Dong <eric.dong@intel.com>
Signed-off-by: Benjamin Doron <benjamin.doron00@gmail.com>